# berkeley-dining-backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Berkeley Dining Menu API",
    "version": "1.1.0"
  },
  "paths": {
...
    "/v1/menu": {
      "get": {
        "summary": "Menu",
        "operationId": "menu_v1_menu_get",
        "parameters": [
          {
            "name": "hall",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Hall"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "meal",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Meal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/menu": {
      "get": {
        "summary": "Menu",
        "operationId": "menu_menu_get",
        "parameters": [
          {
            "name": "hall",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Hall"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "meal",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Meal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Hall": {
        "type": "string",
        "enum": [
          "crossroads",
          "cafe-3",
          "foothill",
          "clark-kerr"
        ],
        "title": "Hall"
      },
      "Item": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories"
          },
          "serving": {
            "$ref": "#/components/schemas/Serving"
          },
          "macros": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Macros"
              },
              {
                "type": "null"
              }
            ]
          },
          "nutrition_status": {
            "type": "string",
            "enum": [
              "published",
              "unapproved",
              "missing"
            ],
            "title": "Nutrition Status"
          },
          "reference_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Image Url"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "categories",
          "serving",
          "macros",
          "nutrition_status"
        ],
        "title": "Item"
      },
      "Macros": {
        "properties": {
          "calories_kcal": {
            "type": "number",
            "minimum": 0.0,
            "title": "Calories Kcal"
          },
          "protein_g": {
            "type": "number",
            "minimum": 0.0,
            "title": "Protein G"
          },
          "carbs_g": {
            "type": "number",
            "minimum": 0.0,
            "title": "Carbs G"
          },
          "fat_g": {
            "type": "number",
            "minimum": 0.0,
            "title": "Fat G"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "calories_kcal",
          "protein_g",
          "carbs_g",
          "fat_g"
        ],
        "title": "Macros"
      },
      "Meal": {
        "type": "string",
        "enum": [
          "breakfast",
          "lunch",
          "dinner",
          "late-night",
          "brunch"
        ],
        "title": "Meal"
      },
      "MenuResponse": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "hall": {
            "$ref": "#/components/schemas/Hall"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "meal": {
            "$ref": "#/components/schemas/Meal"
          },
          "status": {
            "type": "string",
            "enum": [
              "published",
              "not_published"
            ],
            "title": "Status"
          },
          "source_meal_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Meal Names"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Item"
            },
            "type": "array",
            "title": "Items"
          },
          "revision": {
            "type": "string",
            "title": "Revision"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "hall",
          "date",
          "meal",
          "status",
          "source_meal_names",
          "items",
          "revision",
          "fetched_at",
          "expires_at",
          "source_url",
          "source_sha256"
        ],
        "title": "MenuResponse"
      },
      "Serving": {
        "properties": {
          "quantity": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "weight_g": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight G"
          },
          "weight_basis": {
            "type": "string",
            "enum": [
              "source_mass_unit",
              "unknown"
            ],
            "title": "Weight Basis"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "quantity",
          "unit",
          "weight_basis"
        ],
        "title": "Serving"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}